requirements.txt
beautifulsoup4
requests
pyyaml
tornado
ptyprocess